refactor(changelog): extract compareVersionDesc to src/util/semver
Move the version comparison function out of util/changelog.ts into its
own util/semver.ts so both the runtime page code and the upcoming
build-time generator can import one authoritative copy. Re-export it
from util/changelog to preserve existing import paths. Add a dedicated
test suite in test/changelog/semver.test.ts covering numeric sort,
prerelease ranking, and leading-v tolerance.
